map search

+++ src/components/Account.js
@@ -10,6 +10,8 @@ export default class Account extends React.Component {
   }
 
   deleteProfile = () => {
+    if (!confirm('Are your sure? This will delete you data forever')) return;
+
     db.deleteProfile()
     .then(() => this.props.history.push('/'))
     .catch(console.error);
@@ -21,8 +23,12 @@ export default class Account extends React.Component {
     return (
       <div className="container">
         <h1 className="is-size-1">Account</h1>
-        <pre>{user.displayName}</pre>
-        <button onClick={this.deleteProfile} className="button is-danger">Delete Profile</button>
+        <label className="label">Display Name</label>
+        <p>{user.displayName}</p>
+        <h2 className="is-size-3">History</h2>
+        <h2 className="is-size-3">Danger Zone</h2>
+        <br/>
+        <button onClick={this.deleteProfile} className="button is-danger">Delete your Account</button>
       </div>
     );
   }
